modules/cloud_access_scanner.py
#!/usr/bin/env python3
import boto3
from azure.identity import DefaultAzureCredential
from google.cloud import iam_v1
import logging
from typing import List, Dict, Any

class CloudAccessScanner:
    def __init__(self, config: Dict):
        """
        Initialize Cloud Access Scanner with advanced security checks
        
        :param config: Configuration dictionary
        """
        self.config = config
        self.logger = logging.getLogger(__name__)
        
        # Security thresholds for cloud environments
        self.security_thresholds = {
            'max_privileged_users': 5,
            'max_service_accounts': 10,
            'max_external_collaborators': 3
        }

    def _scan_aws_access_controls(self) -> List[Dict[str, Any]]:
        """
        Comprehensive AWS IAM access control scanning
        
        :return: List of AWS IAM security findings
        """
        results = []
        try:
            # Initialize AWS IAM client
            iam_client = boto3.client('iam')
            
            # Check root account usage
            root_usage = iam_client.get_account_summary()
            results.append({
                'control_id': 'AC-2(1)',
                'description': 'AWS Root Account Usage',
                'compliant': root_usage['SummaryMap'].get('AccountMFAEnabled', 0) > 0,
                'details': {
                    'root_mfa_enabled': root_usage['SummaryMap'].get('AccountMFAEnabled', 0) > 0
                },
                'remediation': 'Enable MFA for root account and avoid direct root usage'
            })

            # Check IAM users
            users = iam_client.list_users()
            privileged_users = [
                user for user in users['Users'] 
                if any(policy.startswith('arn:aws:iam::aws:policy/AdministratorAccess') 
                       for policy in iam_client.list_attached_user_policies(UserName=user['UserName'])['AttachedPolicies'])
            ]

            results.append({
                'control_id': 'AC-3(7)',
                'description': 'AWS Privileged User Accounts',
                'compliant': len(privileged_users) <= self.security_thresholds['max_privileged_users'],
                'details': {
                    'privileged_users': [user['UserName'] for user in privileged_users]
                },
                'remediation': f'Limit privileged users to {self.security_thresholds["max_privileged_users"]} or fewer'
            })

        except Exception as e:
            self.logger.error(f"AWS Access Control Scan Error: {e}")
            results.append({
                'control_id': 'AC-AWS-001',
                'description': 'AWS Access Control Scan',
                'compliant': False,
                'remediation': f'Error scanning AWS access controls: {str(e)}'
            })

        return results

    def _scan_azure_access_controls(self) -> List[Dict[str, Any]]:
        """
        Comprehensive Azure Active Directory access control scanning
        
        :return: List of Azure AD security findings
        """
        results = []
        try:
            # Use DefaultAzureCredential for secure authentication
            credential = DefaultAzureCredential()
            
            # Azure AD management requires specific Azure SDK methods
            # Note: This is a placeholder and would need actual Azure SDK implementation
            results.append({
                'control_id': 'AC-2(3)',
                'description': 'Azure AD User Account Management',
                'compliant': True,  # Placeholder
                'details': {
                    'global_admins': [],  # Placeholder
                    'external_users': []  # Placeholder
                },
                'remediation': 'Review Azure AD user accounts and global administrator assignments'
            })

        except Exception as e:
            self.logger.error(f"Azure Access Control Scan Error: {e}")
            results.append({
                'control_id': 'AC-AZURE-001',
                'description': 'Azure Access Control Scan',
                'compliant': False,
                'remediation': f'Error scanning Azure access controls: {str(e)}'
            })

        return results

    def _scan_gcp_access_controls(self) -> List[Dict[str, Any]]:
        """
        Comprehensive Google Cloud Platform IAM access control scanning
        
        :return: List of GCP IAM security findings
        """
        results = []
        try:
            # Initialize GCP IAM client
            client = iam_v1.IAMClient()
            
            # Placeholder for GCP IAM scanning
            results.append({
                'control_id': 'AC-3(5)',
                'description': 'GCP Service Account Management',
                'compliant': True,  # Placeholder
                'details': {
                    'service_accounts': [],  # Placeholder
                    'external_collaborators': []  # Placeholder
                },
                'remediation': 'Review GCP service account permissions and external collaborators'
            })

        except Exception as e:
            self.logger.error(f"GCP Access Control Scan Error: {e}")
            results.append({
                'control_id': 'AC-GCP-001',
                'description': 'GCP Access Control Scan',
                'compliant': False,
                'remediation': f'Error scanning GCP access controls: {str(e)}'
            })

        return results

    def scan_cloud_access_controls(self) -> List[Dict[str, Any]]:
        """
        Perform comprehensive cloud access control scans
        
        :return: List of cloud access control findings
        """
        results = []
        
        # Cloud provider scanning configuration
        cloud_providers = {
            'aws': self._scan_aws_access_controls,
            'azure': self._scan_azure_access_controls,
            'gcp': self._scan_gcp_access_controls
        }
        
        # Scan enabled cloud providers
        for provider, scan_method in cloud_providers.items():
            if self.config.get(provider, {}).get('enabled', False):
                try:
                    results.extend(scan_method())
                except Exception as e:
                    self.logger.error(f"{provider.upper()} Cloud Access Scan Error: {e}")
                    results.append({
                        'control_id': f'AC-{provider.upper()}-ERROR',
                        'description': f'{provider.upper()} Cloud Access Scan Failure',
                        'compliant': False,
                        'remediation': f'Error scanning {provider.upper()} access controls: {str(e)}'
                    })
        
        return results
